@@ -1,30 +1,33 @@
 {
   "name": "app",
-  "version": "0.1.0",
+  "version": "2.0.0",
   "private": true,
   "dependencies": {
-    "@types/jest": "24.0.18",
+    "@types/jest": "29.5.8",
     "@types/lodash": "^4.14.141",
-    "@types/node": "^12.7.8",
-    "@types/react": "^16.9.4",
-    "@types/react-dom": "^16.9.1",
-    "axios": "^0.19.0",
+    "@types/node": "^20.9.1",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "axios": "^1.6.2",
+    "bootstrap": "^5.3.2",
     "lodash": "^4.17.15",
     "moment": "^2.24.0",
-    "node-sass": "^4.12.0",
+    "node-sass": "^9.0.0",
     "psl": "^1.4.0",
-    "react": "^16.10.1",
-    "react-dom": "^16.10.1",
-    "react-helmet": "^5.2.1",
-    "react-loader-spinner": "^3.1.4",
-    "react-quill": "^1.3.3",
-    "react-redux": "^7.1.1",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.1.2",
-    "react-toastify": "^5.4.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.9.1",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
+    "react-icons": "^4.12.0",
+    "react-loader-spinner": "^5.4.5",
+    "react-quill": "^2.0.0",
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.19.0",
+    "react-scripts": "5.0.1",
+    "react-toastify": "^9.1.3",
     "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
-    "typescript": "^3.6.3"
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -49,9 +52,9 @@
   },
   "devDependencies": {
     "@types/moment": "^2.13.0",
-    "@types/react-helmet": "^5.0.11",
+    "@types/react-helmet": "^6.1.9",
     "@types/react-redux": "^7.1.4",
     "@types/react-router-dom": "^5.1.0",
-    "prettier": "^2.0.4"
+    "prettier": "^3.1.0"
   }
 }
