@@ -1,6 +1,6 @@
 
 import React from 'react'
-import { Redirect, Route } from 'react-router-dom'
+import { Navigate, Route } from 'react-router-dom'
 import { UsersState } from '../../../modules/users/redux/states';
 //@ts-ignore
 import { connect } from "react-redux";
@@ -9,25 +9,21 @@ import * as usersOperators from '../../../modules/users/redux/operators'
 
 interface AuthenticatedRouteProps {
   users: UsersState;
-  component: any;
+  component: React.ReactElement;
   path: any;
 }
 
-const AuthenticatedRoute: React.FC<AuthenticatedRouteProps> = ({ users, component: Component, ...rest }) => {
+const AuthenticatedRoute: React.FC<AuthenticatedRouteProps> = ({ users, component, ...rest }) => {
   // Add your own authentication on the below line.
   const isLoggedIn = users.isAuthenticated;
 
+  if (!isLoggedIn) {
+    // Redirect to the login page if the user is not authenticated
+    return <Navigate to={'/'} />;
+  }
+
   return (
-    <Route
-      {...rest}
-      render={props =>
-        isLoggedIn ? (
-          <Component {...props} />
-        ) : (
-          <Redirect to={{ pathname: '/', state: { from: props.location } }} />
-        )
-      }
-    />
+    <Route {...rest} element={component} />
   )
 }
 
