@@ -11,8 +11,20 @@ interface PostRowProps extends Post {
   isLoggedIn: boolean;
 }
 
+function isPostFromToday(date: string | Date): boolean {
+  const currentDate = new Date();
+  const postDate = typeof date === "string" ? new Date(date) : date;
+  let isSameYear = currentDate.getFullYear() == postDate.getFullYear()
+  let isSameMonth = currentDate.getMonth() == postDate.getMonth()
+  let isSameDay = currentDate.getDate() == postDate.getDate()
+  return isSameYear && isSameMonth && isSameDay;
+}
+
 const PostRow: React.FC<PostRowProps> = (props) => (
-  <div className="post-row">
+  <div
+    className="post-row"
+    style={{ backgroundColor: isPostFromToday(props.createdAt) ? "red" : "" }}
+  >
     <Points
       onUpvoteClicked={() => props.onUpvoteClicked()}
       onDownvoteClicked={() => props.onDownvoteClicked()}
