@@ -1,6 +1,7 @@
 
 import React from 'react'
 import { Button } from '../../../../../shared/components/button'
+import { Link } from 'react-router-dom';
 
 interface ProfileButtonProps {
   isLoggedIn: boolean;
@@ -11,7 +12,7 @@ interface ProfileButtonProps {
 const ProfileButton: React.FC<ProfileButtonProps> = (props) => {
   return props.isLoggedIn ? (
     <Button 
-      text={<span>{`${props.username} / `}{<u onClick={props.onLogout}>logout</u>}</span>}
+      text={<span><Link to={`/member/${props.username}`} style={{color: 'white'}}>{props.username}</Link> / {<u onClick={props.onLogout}>logout</u>}</span>}
       onClick={() => {}}
     />
   ) : (
