@@ -5,9 +5,9 @@ import * as actions from "./actions";
 import { UsersAction } from "./actionCreators";
 import { ReduxUtils } from "../../../shared/utils/ReduxUtils";
 
-export default function users (state: UsersState = states,
+export default function users(state: UsersState = states,
   action: UsersAction
-) : UsersState {
+): UsersState {
   switch (action.type as actions.UsersActionType) {
     case actions.GETTING_USER_PROFILE:
       return {
@@ -26,7 +26,64 @@ export default function users (state: UsersState = states,
         ...state,
         ...ReduxUtils.reportEventStatus("isFetchingUser", false)
       };
-    
+
+    case actions.GETTING_USER_STATISTICS:
+      return {
+        ...state,
+        ...ReduxUtils.reportEventStatus("isFetchingUserStats"),
+        error: ''
+      };
+    case actions.GETTING_USER_STATISTICS_SUCCESS:
+      return {
+        ...state,
+        ...ReduxUtils.reportEventStatus("isFetchingUserStats", true),
+        userStats: action.userStats
+      };
+    case actions.GETTING_USER_STATISTICS_FAILURE:
+      return {
+        ...state,
+        ...ReduxUtils.reportEventStatus("isFetchingUserStats", false),
+        error: action.error
+      };
+
+    case actions.GETTING_USER_STATISTICS_WITH_BEST_SCORE:
+      return {
+        ...state,
+        ...ReduxUtils.reportEventStatus("isFetchingUserStatsWithBestScore"),
+        error: ''
+      };
+    case actions.GETTING_USER_STATISTICS_WITH_BEST_SCORE_SUCCESS:
+      return {
+        ...state,
+        ...ReduxUtils.reportEventStatus("isFetchingUserStatsWithBestScore", true),
+        userStatsWithBestScore: action.userStatsWithBestScore
+      };
+    case actions.GETTING_USER_STATISTICS_WITH_BEST_SCORE_FAILURE:
+      return {
+        ...state,
+        ...ReduxUtils.reportEventStatus("isFetchingUserStatsWithBestScore", false),
+        error: action.error
+      };
+
+    case actions.GETTING_USER_STATISTICS_WITH_MOST_POSTS:
+      return {
+        ...state,
+        ...ReduxUtils.reportEventStatus("isFetchingUserStatsWithMostPosts"),
+        error: ''
+      };
+    case actions.GETTING_USER_STATISTICS_WITH_MOST_POSTS_SUCCESS:
+      return {
+        ...state,
+        ...ReduxUtils.reportEventStatus("isFetchingUserStatsWithMostPosts", true),
+        userStatsWithMostPosts: action.userStatsWithMostPosts
+      };
+    case actions.GETTING_USER_STATISTICS_WITH_MOST_POSTS_FAILURE:
+      return {
+        ...state,
+        ...ReduxUtils.reportEventStatus("isFetchingUserStatsWithMostPosts", false),
+        error: action.error
+      };
+
     case actions.LOGGING_IN:
       return {
         ...state,
@@ -64,7 +121,7 @@ export default function users (state: UsersState = states,
         ...ReduxUtils.reportEventStatus('isLoggingOut', false),
         error: action.error
       }
-    
+
     case actions.CREATING_USER:
       return {
         ...state,
@@ -75,13 +132,13 @@ export default function users (state: UsersState = states,
       return {
         ...state,
         ...ReduxUtils.reportEventStatus('isCreatingUser', true),
-      }     
+      }
     case actions.CREATING_USER_FAILURE:
       return {
         ...state,
         ...ReduxUtils.reportEventStatus('isCreatingUser', false),
         error: action.error
-      }       
+      }
     default:
       return state;
   }
