@@ -24,12 +24,11 @@ import { SubmitButton } from '../shared/components/button';
 import { TextUtil } from '../shared/utils/TextUtil';
 import { FullPageLoader } from '../shared/components/loader';
 import withVoting from '../modules/forum/hocs/withVoting';
-import { Points } from '../modules/forum/components/posts/points';
+//import { Points } from '../modules/forum/components/posts/points';
 
 interface DiscussionPageProps extends usersOperators.IUserOperators, forumOperators.IForumOperations {
   users: UsersState;
   forum: ForumState;
-  history: any;
 }
 
 interface DiscussionState {
@@ -133,6 +132,7 @@ class DiscussionPage extends React.Component<DiscussionPageProps, DiscussionStat
   render () {
     const post = this.props.forum.post as Post;
     const comments = this.props.forum.comments;
+    console.log(this.props);
 
     return (
       <Layout>
