@@ -1,5 +1,5 @@
 import React from 'react'
-import arrow from '../assets/arrow.svg'
+//import arrow from '../assets/arrow.svg'
 import { Link } from 'react-router-dom'
 import "../styles/BackNavigation.sass"
 
