@@ -1,11 +1,10 @@
 
 import React from 'react';
 //@ts-ignore
-import RLoader from 'react-loader-spinner'
+import { Rings } from 'react-loader-spinner'
 
 const Loader = () => (
-  <RLoader
-    type="Rings"
+  <Rings
     color="#6a69ff"
     height={100}
     width={100}
