@@ -5,7 +5,7 @@
 */
 import { Logger } from "tslog";
 
-import ConfigHandler from "../../config/ConfigHandler";
+import ConfigHandler from "../../config/configHandler";
 import  { RestClient }  from "../../restClient/RestClient";
 
 export abstract class AEndpoint {
