@@ -1,3 +1,3 @@
 // API tests for global endpoints
-import './users.partial-test.ts';
-import './forum.partial-test.ts';
+import './users.partial-test';
+import './forum.partial-test';
