@@ -1,5 +1,5 @@
 
-import uuid from 'uuid/v4';
+import { v4 as uuid } from 'uuid';
 import { Identifier } from './Identifier'
 
 export class UniqueEntityID extends Identifier<string | number>{
