@@ -1,7 +1,7 @@
 
 import { getRecentPosts } from '../../../../modules/forum/useCases/post/getRecentPosts'
 import { PostDetailsMap } from '../../../../modules/forum/mappers/postDetailsMap'
-import { GraphQLDateTime } from 'graphql-iso-date';
+import { DateTimeResolver } from 'graphql-scalars'
 import { memberRepo } from '../../../../modules/forum/repos'
 import { MemberDetailsMap } from '../../../../modules/forum/mappers/memberDetailsMap'
 import { ApolloServer, gql } from 'apollo-server-express';
@@ -24,7 +24,7 @@ const server = new ApolloServer({
     UserTypeDefs
   ],
   resolvers: {
-    DateTime: GraphQLDateTime,
+    DateTime: DateTimeResolver,
     Post: {
       memberPostedBy: async (post, args, context) => {
         const memberDetails = await memberRepo.getMemberDetailsByPostLinkOrSlug(post.slug);
